# apps/web/openapi/gateway.json
{
  "components": {
    "schemas": {
      "AgentApprovalDecision": {
        "additionalProperties": false,
        "properties": {
          "clientVersion": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clientversion"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "title": "AgentApprovalDecision",
        "type": "object"
      },
...
      "AgentMessageResponse": {
        "additionalProperties": false,
        "properties": {
          "completedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completedat"
          },
          "content": {
            "$ref": "#/components/schemas/AgentMessageContent"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "providerItemId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provideritemid"
          },
          "role": {
            "$ref": "#/components/schemas/AgentMessageRole"
          },
          "runId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runid"
          },
          "sequence": {
            "minimum": 1.0,
            "title": "Sequence",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/AgentMessageStatus"
          },
          "threadId": {
            "title": "Threadid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "threadId",
          "sequence",
          "role",
          "content",
          "status",
          "createdAt"
        ],
        "title": "AgentMessageResponse",
        "type": "object"
      },
...
      "AgentRunCreate": {
        "additionalProperties": false,
        "properties": {
          "assetIds": {
            "items": {
              "maxLength": 64,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 4,
            "title": "Assetids",
            "type": "array"
          },
          "conversationId": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Conversationid",
            "type": "string"
          },
          "draftSnapshot": {
            "$ref": "#/components/schemas/GenerationDraft"
          },
          "message": {
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "mode": {
            "default": "assist",
            "enum": [
              "assist",
              "image-analysis",
              "storyboard"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "conversationId",
          "message",
          "draftSnapshot"
        ],
        "title": "AgentRunCreate",
        "type": "object"
      },
      "AgentRunResponse": {
        "additionalProperties": false,
        "properties": {
          "conversationId": {
            "title": "Conversationid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormessage"
          },
          "finishedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finishedat"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inputSnapshot": {
            "additionalProperties": true,
            "title": "Inputsnapshot",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "outputMessageId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outputmessageid"
          },
          "promptVersion": {
            "title": "Promptversion",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "providerResponseId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerresponseid"
          },
          "reasoningEffort": {
            "title": "Reasoningeffort",
            "type": "string"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "status": {
            "$ref": "#/components/schemas/AgentRunStatus"
          },
          "threadId": {
            "title": "Threadid",
            "type": "string"
          },
          "toolsetVersion": {
            "title": "Toolsetversion",
            "type": "string"
          },
          "transport": {
            "enum": [
              "sse",
              "websocket"
            ],
            "title": "Transport",
            "type": "string"
          },
          "usage": {
            "$ref": "#/components/schemas/AgentUsage"
          }
        },
        "required": [
          "id",
          "threadId",
          "conversationId",
          "status",
          "model",
          "provider",
          "transport",
          "reasoningEffort",
          "promptVersion",
          "toolsetVersion",
          "inputSnapshot",
          "createdAt"
        ],
        "title": "AgentRunResponse",
        "type": "object"
      },
...
      "AgentToolDecisionResponse": {
        "additionalProperties": false,
        "properties": {
          "approval": {
            "$ref": "#/components/schemas/AgentApprovalResponse"
          },
          "run": {
            "$ref": "#/components/schemas/AgentRunResponse"
          },
          "toolCall": {
            "$ref": "#/components/schemas/AgentToolCallResponse"
          }
        },
        "required": [
          "toolCall",
          "approval",
          "run"
        ],
        "title": "AgentToolDecisionResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PromptEnhanceRequest": {
        "additionalProperties": false,
        "properties": {
          "negativePrompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negativeprompt"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "PromptEnhanceRequest",
        "type": "object"
      },
      "PromptEnhanceResponse": {
        "additionalProperties": false,
        "properties": {
          "negativePrompt": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Negativeprompt"
          },
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "PromptEnhanceResponse",
        "type": "object"
      },
...
      "TitleSummarizeRequest": {
        "additionalProperties": false,
        "properties": {
          "prompt": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          }
        },
        "required": [
          "prompt"
        ],
        "title": "TitleSummarizeRequest",
        "type": "object"
      },
      "TitleSummarizeResponse": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "TitleSummarizeResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Oneiroi Studio Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/agent/prompt-enhance": {
      "post": {
        "operationId": "prompt_enhance_v1_agent_prompt_enhance_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptEnhanceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptEnhanceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Prompt Enhance",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/runs": {
      "post": {
        "operationId": "create_run_v1_agent_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentRunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/runs/{run_id}": {
      "get": {
        "operationId": "get_run_v1_agent_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_v1_agent_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/runs/{run_id}/events": {
      "get": {
        "operationId": "run_events_v1_agent_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Events",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/threads/{thread_id}/messages": {
      "get": {
        "operationId": "list_messages_v1_agent_threads__thread_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentMessageResponse"
                  },
                  "title": "Response List Messages V1 Agent Threads  Thread Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Messages",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/title-summarize": {
      "post": {
        "operationId": "title_summarize_v1_agent_title_summarize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TitleSummarizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TitleSummarizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Title Summarize",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/tool-calls/{tool_call_id}/approve": {
      "post": {
        "operationId": "approve_tool_call_v1_agent_tool_calls__tool_call_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_call_id",
            "required": true,
            "schema": {
              "title": "Tool Call Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentApprovalDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentToolDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Tool Call",
        "tags": [
          "agent"
        ]
      }
    },
    "/v1/agent/tool-calls/{tool_call_id}/reject": {
      "post": {
        "operationId": "reject_tool_call_v1_agent_tool_calls__tool_call_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_call_id",
            "required": true,
            "schema": {
              "title": "Tool Call Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentApprovalDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentToolDecisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Tool Call",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/v1/conversations/{conversation_id}/agent/thread": {
      "get": {
        "operationId": "get_thread_v1_conversations__conversation_id__agent_thread_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Oneiroi-User",
            "required": false,
            "schema": {
              "default": "demo-user",
              "title": "X-Oneiroi-User",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentThreadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Thread",
        "tags": [
          "agent"
        ]
      }
    },
...
  }
}